Git: Check for SSH key presence before trying to clone
Many users are having trouble cloning, and error handling is very poor.
This is a first step towards improving it.
